+++ examples/data-sources/upcloud_load_balancer_dns_challenge_domain/data-source.tf
@@ -1,0 +1,3 @@
+data "upcloud_load_balancer_dns_challenge_domain" "this" {}
+
+// Configure _acme-challenge by creating "_acme-challenge.objects IN CNAME ${data.upcloud_load_balancer_dns_challenge_domain.this.domain}" DNS record.